+++ src/app/v2/FeatureHubV2.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function FeatureHubV2() {
+  return <div>v2 hub</div>;
+}